AC_Autorotation: Fix data type for time_on_ground

+++ libraries/AC_Autorotation/AC_Autorotation.h
@@ -155,7 +155,7 @@ private:
     float _curr_sink_deriv[20];      // Array for storing sink rate derivatives
     bool  _flare_complete;           // Flare completed
     bool  _flare_update_check;       // Check for flare altitude updating
-    float _time_on_ground;           // Time elapsed after touch down
+    uint32_t _time_on_ground;        // Time elapsed after touch down
 
     LowPassFilterFloat _accel_target_filter; // acceleration target filter
